Use information from the dnsmasq hosts file to call dhcp_release
Certain situations can cause the DHCP agent's local cache to get out
of sync with the leases held internally by dnsmasq.  This method of
detecting when to call dhcp_release is idempotent and not dependent on
the cache.  It is more robust.

Hyunsun: 'dhcp_release: CommandFilter, dhcp_release, root' is required
in rootwrap.d/dhcp.filter

Change-Id: I4eafd9cfb94a77a2f0229f89de5483dad23725cf
Closes-Bug: #1250644

Conflicts:

	neutron/agent/linux/dhcp.py
	neutron/tests/unit/test_linux_dhcp.py