Return freed IP to the pool right after the release
Quantum waits for the freed IP to be expired (dhcp lease) to
return to the pool. This results in successive VIP move failure.
With this fix, freed IP will be available as soon as it is released
so that the VIP move can be done regardless of dhcp_lease_duration.